pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "device-ingestion-service"
version = "0.1.0"
description = "Wave 3 MQTT ingestion extraction service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "paho-mqtt>=2.1.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "uvicorn>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "mypy>=1.13.0",
  "pytest>=8.3.0",
  "ruff>=0.7.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/device_ingestion_service"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "postgres_integration: requires a live PostgreSQL instance"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
mypy_path = "src"

[tool.coverage.run]
source = ["device_ingestion_service"]
branch = true
